pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simulaiz"
version = "0.1.0"
description = "Multi-agent simulation toolkit."
authors = [{ name = "SimulAiz Maintainers" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.119.0",
    "uvicorn[standard]==0.37.0",
    "python-multipart>=0.0.6",
    "PyJWT==2.10.1",
    "livekit==1.0.11",
    "numpy==1.26.4",
    "scipy==1.12.0",
    "webrtcvad-wheels==2.0.14",
    "faster-whisper==1.0.3",
    "TTS==0.22.0",
    "opencv-python-headless==4.8.1.78",
    "diffusers==0.31.0",
    "transformers==4.46.2",
    "accelerate==1.1.1",
    "safetensors==0.4.3",
    "Pillow==10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
    "pre-commit>=3.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]

 
